Move EventDisplay to Home views folder.
