+++ frontend/src/views/Home/EventDisplay.vue
